Add handling for Type C applications
